+++ R/helpers_miscellaneous.R
@@ -9,23 +9,31 @@
 #' @param stat.title A character describing the test being run, which will be
 #'   added as a prefix in the subtitle. The default is `NULL`. An example of a
 #'   `stat.title` argument will be something like `"Student's t-test: "`.
+#' @param stats.df A dataframe containing the following columns:
+#' \itemize{
+#'   \item *statistic*: the numeric value of a statistic.
+#'   \item *parameter*: the numeric value of a parameter being modeled (often
+#' degrees of freedom for the test); note that if `no.parameters = 0L` (e.g.,
+#' for non-parametric tests), this column will be irrelevant.
+#'   \item *parameter1*, *parameter2* relevant only if the statistic in question
+#' has two degrees of freedom (e.g., anova).
+#'   \item *p.value* the two-sided *p*-value associated with the observed
+#' statistic.
+#' }
 #' @param statistic.text A character that specifies the relevant test statistic.
 #'   For example, for tests with *t*-statistic, `statistic.text = "t"`. If you
 #'   want to use plotmath, you will have to quote the argument (e.g.,
 #'   `quote(italic("t"))`).
-#' @param statistic The numeric value of a statistic.
-#' @param parameter The numeric value of a parameter being modeled (often
-#'   degrees of freedom for the test). Default is `NULL` to accommodate
-#'   non-parametric tests.
-#' @param parameter2 Relevant only if the statistic in question has two degrees
-#'   of freedom (default: `NULL`).
-#' @param p.value The two-sided p-value associated with the observed statistic.
+#' @param effsize.df A dataframe containing the following columns:
+#' \itemize{
+#'   \item *estimate*: estimated value of the effect size.
+#'   \item *conf.low*:  lower bound for effect size estimate.
+#'   \item *conf.high*: upper bound for effect size estimate.
+#' }
 #' @param effsize.text A character that specifies the relevant effect size.
 #'   For example, for Cohen's *d* statistic, `effsize.text = "d"`. If you
 #'   want to use plotmath, you will have to quote the argument (e.g.,
 #'   `quote(italic("d"))`).
-#' @param effsize.estimate,effsize.LL,effsize.UL The estimated value of the
-#'   effect size, its lower bound, and its upper.
 #' @param k Number of digits after decimal point (should be an integer)
 #'   (Default: `k = 2`).
 #' @param k.parameter,k.parameter2 Number of decimal places to display for the
@@ -36,22 +44,33 @@
 #'   `quote(italic("n")["pairs"])`, while for independent subjects design this
 #'   can be `quote(italic("n")["obs"])`. If `NULL`, defaults to generic
 #'   `quote(italic("n"))`.
+#' @param ... Currently ignored.
 #' @inheritParams t1way_ci
 #'
 #' @examples
 #' set.seed(123)
 #'
+#' # creating a dataframe with stats results
+#' stats_df <- cbind.data.frame(
+#'   statistic = 5.494,
+#'   parameter = 29.234,
+#'   p.value = 0.00001
+#' )
+#'
+#' # creating a dataframe with effect size results
+#' effsize_df <- cbind.data.frame(
+#'   estimate = -1.980,
+#'   conf.low = -2.873,
+#'   conf.high = -1.088
+#' )
+#'
 #' # subtitle for *t*-statistic with Cohen's *d* as effect size
 #' statsExpressions::expr_template(
 #'   no.parameters = 1L,
+#'   stats.df = stats_df,
+#'   effsize.df = effsize_df,
 #'   statistic.text = quote(italic("t")),
-#'   statistic = 5.494,
-#'   parameter = 29.234,
-#'   p.value = 0.00001,
 #'   effsize.text = quote(italic("d")),
-#'   effsize.estimate = -1.980,
-#'   effsize.LL = -2.873,
-#'   effsize.UL = -1.088,
 #'   n = 32L,
 #'   conf.level = 0.95,
 #'   k = 3L,
@@ -63,23 +82,26 @@
 expr_template <- function(no.parameters,
                           stat.title = NULL,
                           statistic.text,
-                          statistic,
-                          parameter = NULL,
-                          parameter2 = NULL,
-                          p.value,
+                          stats.df,
                           effsize.text,
-                          effsize.estimate,
-                          effsize.LL,
-                          effsize.UL,
+                          effsize.df,
                           n,
                           conf.level = 0.95,
                           k = 2L,
                           k.parameter = 0L,
                           k.parameter2 = 0L,
-                          n.text = NULL) {
+                          n.text = NULL,
+                          ...) {
 
   # if sample size nature is not specified, use generic n
   if (rlang::is_null(n.text)) n.text <- quote(italic("n"))
+
+  # extracting the common values
+  statistic <- stats.df$statistic[[1]]
+  p.value <- stats.df$p.value[[1]]
+  effsize.estimate <- effsize.df$estimate[[1]]
+  effsize.LL <- effsize.df$conf.low[[1]]
+  effsize.UL <- effsize.df$conf.high[[1]]
 
   # ------------------ statistic with 0 degrees of freedom --------------------
 
@@ -131,16 +153,7 @@ expr_template <- function(no.parameters,
 
   if (no.parameters == 1L) {
     # check if parameter is specified
-    if (rlang::is_null(parameter)) {
-      stop(message(cat(
-        crayon::red("Error: "),
-        crayon::blue("For tests with statistic that have two parameters, \n"),
-        crayon::blue("the argument `parameter` **must** be specified.\n"),
-        sep = ""
-      )),
-      call. = FALSE
-      )
-    }
+    parameter <- stats.df$parameter[[1]]
 
     # preparing subtitle
     subtitle <-
@@ -192,18 +205,8 @@ expr_template <- function(no.parameters,
 
   if (no.parameters == 2L) {
     # check if parameters are specified
-    if (rlang::is_null(parameter) || rlang::is_null(parameter2)) {
-      stop(message(cat(
-        crayon::red("Error: "),
-        crayon::blue(
-          "For tests with statistic that have two parameters, \n",
-          "the arguments `parameter` and `parameter2` **must** be specified.\n"
-        ),
-        sep = ""
-      )),
-      call. = FALSE
-      )
-    }
+    parameter <- stats.df$parameter1[[1]]
+    parameter2 <- stats.df$parameter2[[1]]
 
     # preparing subtitle
     subtitle <-
@@ -259,16 +262,28 @@ expr_template <- function(no.parameters,
 }
 
 #' @noRd
+#'
+#' @importFrom dplyr rename_all recode
+#'
 #' @keywords internal
 
-rcompanion_cleaner <- function(object, estimate.col) {
-  object %>%
-    tibble::as_tibble(x = .) %>%
-    dplyr::rename(
-      .data = .,
-      estimate = estimate.col,
-      conf.low = lower.ci,
-      conf.high = upper.ci
+rcompanion_cleaner <- function(object) {
+  # if a list, extract the first component containing results
+  if (inherits(object, "list")) object <- object[[1]]
+
+  # rename columns uniformly
+  tibble::as_tibble(object) %>%
+    dplyr::rename_all(
+      .tbl = .,
+      .funs = dplyr::recode,
+      epsilon.squared = "estimate",
+      r = "estimate",
+      rho = "estimate",
+      W = "estimate",
+      Cramer.V = "estimate",
+      Value = "estimate",
+      lower.ci = "conf.low",
+      upper.ci = "conf.high"
     )
 }